+++ sdk/core/azure-core-management/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.0-beta.1 (Unreleased)
 
+- Fixed long-running operation, PUT method, response 200 and Azure-AsyncOperation.
 
 ## 1.1.0 (2021-01-11)